requirements.txt
absl-py==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
cachetools==5.3.3
catalogue==1.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==43.0.1
cvzone==1.5.6
cycler==0.11.0
cymem==2.0.8
decorator==5.1.1
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.33.3
fsspec==2024.9.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.26.0
idna==3.7
imbalanced-learn==0.12.3
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
mysql==0.0.3
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.16.0
networkx==3.4.1
nltk==3.9.1
numpy==1.24.4
oauthlib==3.2.2
opencv-contrib-python==4.5.5.64
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==21.3
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20240706
Pillow==9.1.0
plac==1.1.3
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
pyresparser==1.0.6
pyrsistent==0.20.0
PySide2==5.15.2.1
python-dateutil==2.8.2
python-multipart==0.0.12
python-version==0.0.2
pytz==2022.1
PyYAML==6.0.2
qibullet==1.4.5
referencing==0.35.1
regex==2024.9.11
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.9.0
seaborn==0.11.2
self==2020.12.3
sentence-transformers==3.2.0
shellingham==1.5.4
shiboken2==5.15.2.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.6
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==1.0.7
starlette==0.40.0
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-cpu==2.17.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==7.4.6
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
tornado==6.4.2
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.31.1
wasabi==0.10.1
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
